Continue styling for project
In the CSS file more styling has been implemented to get the
images and text in the correct positions. Incomplete as of now.
Slight adjustment in the HTML by adding a wrapper for the images as
well, in anticipation of it being an appropriate solution
to the correct image placing.