Continue progress on project
More styles implemented in order to get the project closer to the end
result. The arrow down icons have been added to the question.
Uncertainty on whether or not it will be doable the way that they were
implemented when user interactivity is added.
Bottom borders to the various questions as well as hiding the answers
at this time.
Working out issues with proper image positioning.