Edit HTML file and more styles in styles.css
Took the box illustration image out of container and placed it at the
top before the container. This was done to position absolute, and still
be able to see it if it overflows out of the container.

In the CSS, more styles were applied to position the images properly,
before working on the FAQ section. Besides some minor corrections to
positioning, this may be left as is as of now.